requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
transformers==4.35.2
torch==2.6.0
python-multipart==0.0.6
huggingface-hub==0.19.4
pydantic==2.5.2 